Restructure generated proto files
