+++ core/kernel/src/c_interface/mod.rs
@@ -1,3 +1,4 @@
 pub mod interface;
 mod feature_manifest_fbs_mapping;
 mod kernel_failure_fbs_mapping;
+mod kernel_bootstrap_fbs_mapping;
