+++ core/kernel/src/flatbuffers_generated/feature_manifest_generated.rs
