+++ core/kernel/src/flatbuffers_generated/kernel_bootstrap_response_generated.rs
@@ -2,7 +2,7 @@
 
 
 
-use crate::failure_generated::*;
+use crate::flatbuffers_generated::failure_generated::*;
 use std::mem;
 use std::cmp::Ordering;
 
@@ -12,7 +12,7 @@ use self::flatbuffers::{EndianScalar, Follow};
 #[allow(unused_imports, dead_code)]
 pub mod kernel_bootstrap_response {
 
-  use crate::failure_generated::*;
+  use crate::flatbuffers_generated::failure_generated::*;
   use std::mem;
   use std::cmp::Ordering;
