+++ core/kernel/src/flatbuffers_generated/mod.rs
@@ -1,0 +1,8 @@
+#[allow(unused_imports)]
+pub mod feature_manifest_generated;
+#[allow(unused_imports)]
+pub mod feature_manifest_response_generated;
+#[allow(unused_imports)]
+pub mod failure_generated;
+#[allow(unused_imports)]
+pub mod kernel_bootstrap_response_generated;
