+++ core/kernel/src/lib.rs
@@ -1,12 +1,5 @@
-pub mod c_interface;
+mod c_interface;
 mod feature_manifest_parsing;
-pub mod kernel;
-#[allow(unused_imports)]
-mod feature_manifest_generated;
-#[allow(unused_imports)]
-mod feature_manifest_response_generated;
-#[allow(unused_imports)]
-mod failure_generated;
-#[allow(unused_imports)]
-mod kernel_bootstrap_response_generated;
+mod kernel;
 mod kernel_failure;
+mod flatbuffers_generated;
